refactor(client): move signup container code to custom hook
